refactor(redteam): relocate AgentTarget + AgentContext family to evaluatorq.contracts
Move AgentTarget ABC and the AgentContext/ToolInfo/MemoryStoreInfo/
KnowledgeBaseInfo models from redteam internals to the shared
evaluatorq.contracts module so simulation can depend on the target shape
without crossing into the redteam package. Backend, BareTargetBackend, and
validate_agent_target stay in redteam/backends/base.py (Backend is
redteam-specific). redteam/contracts.py keeps re-export shims for the moved
context models; no back-compat re-export of AgentTarget from base.py (clean
break, CHANGELOG'd breaking change). Migrate all 13 importers.

RES-808 PR2/3.
