+++ packages/evaluatorq-py/CHANGELOG.md
@@ -34,6 +34,18 @@ red_team(
 )
 ```
 
+- **`AgentTarget` relocated**: moved from `evaluatorq.redteam.backends.base` to `evaluatorq.contracts`. Importing it from the old path now raises `ImportError`. The `Backend` ABC stays in `evaluatorq.redteam.backends.base`. `AgentContext`, `ToolInfo`, `MemoryStoreInfo`, and `KnowledgeBaseInfo` also moved to `evaluatorq.contracts` (re-exported from `evaluatorq.redteam.contracts` for back-compat).
+
+**Migration:**
+
+```python
+# Before
+from evaluatorq.redteam.backends.base import AgentTarget
+
+# After
+from evaluatorq.contracts import AgentTarget
+```
+
 ### New Features
 
 - **`LLMCallConfig`** — per-role LLM configuration with `model`, `temperature`, `max_tokens`, `timeout_ms`, `extra_kwargs`, and `client` fields
